requirements.txt
fastapi
uvicorn
python-multipart
langchain
langchain-community
langchain-google-genai
langchain-text-splitters
langchain-huggingface
faiss-cpu
sentence-transformers
pypdf
python-dotenv
python-docx
openpyxl
docx2txt
